requirements.txt
boto3==1.3.1
botocore==1.4.39
docutils==0.12
futures==3.0.5
jmespath==0.9.0
pkg-resources==0.0.0
python-dateutil==2.5.3
six==1.10.0
